Moving argv treatment to specific file
